telegram_bot/requirements.txt
python-telegram-bot
pika
python-dotenv